--- src/components/avatarButton/index.js
@@ -1,1 +1,0 @@
-export { default as AvatarButton } from './AvatarButton';
