+++ src/components/button/avatarButton/index.js
@@ -1,0 +1,1 @@
+export { default } from './AvatarButton';
