+++ src/components/button/index.js
@@ -1,0 +1,3 @@
+export { default as AvatarButton } from './avatarButton';
+export { default as CustomButton } from './customButton';
+export { default as RadioButton } from './radioButton';
