+++ src/components/chart/Chart.jsx
@@ -1,6 +1,6 @@
 import chartImg from '@/assets/images/chart.png';
 import loadingSpinner from '@/assets/images/loading-spinner.gif';
-import { CustomButton } from '@/components/customButton';
+import { CustomButton } from '@/components/button';
 import { ENDPOINTS } from '@/constants/api';
 import { requestGet } from '@/utils/api';
 import { useEffect, useState } from 'react';
