--- src/components/customButton/index.js
@@ -1,1 +1,0 @@
-export { default as CustomButton } from './CustomButton';
