+++ src/components/modals/commonModal/index.js
@@ -1,0 +1,1 @@
+export { default } from './Modal';
