+++ src/components/modals/commonModal/modal.styles.js
