+++ src/components/modals/creditChargeModal/CreditChargeModal.jsx
@@ -1,7 +1,6 @@
 import creditWhiteImg from '@/assets/images/credit-white.png';
 import creditImg from '@/assets/images/credit.png';
-import { CustomButton } from '@/components/customButton';
-import { RadioButton } from '@/components/radioButton';
+import { CustomButton, RadioButton } from '@/components/button';
 import { prices } from '@/data/mockData';
 import { useState } from 'react';
 import * as S from './creditChargeModal.styles';
