+++ src/components/modals/creditChargeModal/index.js
@@ -1,1 +1,1 @@
-export { default as CreditChargeModal } from './creditChargeModal';
+export { default } from './creditChargeModal';
