+++ src/components/modals/donationModal/DonationModal.jsx
@@ -1,5 +1,5 @@
 import creditImg from '@/assets/images/credit.png';
-import { CustomButton } from '@/components/customButton';
+import { CustomButton } from '@/components/button';
 import { ENDPOINTS } from '@/constants/api';
 import { requestPut } from '@/utils/api';
 import { useRef, useState } from 'react';
