+++ src/components/modals/donationModal/index.js
@@ -1,1 +1,1 @@
-export { default as DonationModal } from './DonationModal';
+export { default } from './DonationModal';
