+++ src/components/modals/index.js
@@ -1,0 +1,5 @@
+export { default as Modal } from './commonModal';
+export { default as VoteModal } from './voteModal';
+export { default as DonationModal } from './donationModal';
+export { default as CreditChargeModal } from './creditChargeModal';
+export { default as CreditLackModal } from './creditLackModal';
