--- src/components/radioButton/index.js
@@ -1,1 +1,0 @@
-export { default as RadioButton } from './RadioButton';
