+++ src/errorBoundary/apiErrorBoundary/ApiErrorBoundary.jsx
@@ -1,4 +1,4 @@
-import { CustomButton } from '@/components/customButton';
+import { CustomButton } from '@/components/button';
 import { UI_ERRORS } from '@/constants/errors';
 import { STATUS_CODES } from '@/constants/statusCodes';
 import * as S from '@/errorBoundary/styles';
