+++ src/errorBoundary/renderErrorBoundary/RenderErrorBoundary.jsx
@@ -1,4 +1,4 @@
-import { CustomButton } from '@/components/customButton';
+import { CustomButton } from '@/components/button';
 import { UI_ERRORS } from '@/constants/errors';
 import * as S from '@/errorBoundary/styles';
 import { Component } from 'react';
