+++ src/pages/error/notFound/NotFoundPage.jsx
@@ -1,5 +1,5 @@
 import logoImg from '@/assets/images/logo.png';
-import { CustomButton } from '@/components/customButton';
+import { CustomButton } from '@/components/button';
 import { UI_ERRORS } from '@/constants/errors';
 import { STATUS_CODES } from '@/constants/statusCodes';
 import * as S from '@/pages/error/styles';
