+++ src/pages/landing/LandingPage.jsx
@@ -1,4 +1,4 @@
-import { CustomButton } from '@/components/customButton';
+import { CustomButton } from '@/components/button';
 import { motion } from 'motion/react';
 import { useEffect, useRef, useState } from 'react';
 import { Link } from 'react-router-dom';
